rework interface for managing members

+++ app/services/work_package_members/delete_role_service.rb
@@ -1,6 +1,6 @@
-#-- copyright
+# -- copyright
 # OpenProject is an open source project management software.
-# Copyright (C) 2012-2024 the OpenProject GmbH
+# Copyright (C) 2010-2023 the OpenProject GmbH
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License version 3.
@@ -24,40 +24,14 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #
 # See COPYRIGHT and LICENSE files for more details.
-#++
+# ++
 
-class Members::DeleteService < BaseServices::Delete
-  include Members::Concerns::CleanedUp
-
+class WorkPackageMembers::DeleteRoleService < WorkPackageMembers::DeleteService
   def destroy(object)
-    if object.member_roles.where.not(inherited_from: nil).empty?
+    if object.member_roles.where.not("inherited_from IS NULL AND role_id = ?", params[:role_id]).empty?
       super
     else
-      object.member_roles.where(inherited_from: nil).destroy_all
+      object.member_roles.where(inherited_from: nil, role_id: params[:role_id]).destroy_all
     end
-  end
-
-  protected
-
-  def after_perform(service_call)
-    super(service_call).tap do |call|
-      member = call.result
-
-      cleanup_for_group(member)
-      send_notification(member)
-    end
-  end
-
-  def send_notification(member)
-    ::OpenProject::Notifications.send(OpenProject::Events::MEMBER_DESTROYED,
-                                      member:)
-  end
-
-  def cleanup_for_group(member)
-    return unless member.principal.is_a?(Group)
-
-    Groups::CleanupInheritedRolesService
-      .new(member.principal, current_user: user, contract_class: EmptyContract)
-      .call
   end
 end
